+++ tests/vitest.d.ts
@@ -1,9 +1,8 @@
 /* eslint-disable @typescript-eslint/no-empty-object-type */
-/* eslint-disable @typescript-eslint/no-explicit-any */
 import 'vitest';
 
 import { CustomMatcher } from '../src/matcher.ts';
 
 declare module 'vitest' {
-  interface Matchers<T = any> extends CustomMatcher<T> {}
+  interface Matchers<R, T> extends CustomMatcher<R, T> {}
 }